// NhapKhoApp/src/generated/models/Htn1423_nghiasanphamsModel.ts
/*!
 * Copyright (C) Microsoft Corporation. All rights reserved.
 * This file is autogenerated. Do not edit this file directly.
 */
export const Htn1423_nghiasanphamsstatecode = {
  0: 'Active',
  1: 'Inactive'
} as const;
export type Htn1423_nghiasanphamsstatecode = keyof typeof Htn1423_nghiasanphamsstatecode;
export const Htn1423_nghiasanphamsstatuscode = {
  1: 'Active',
  2: 'Inactive'
} as const;
export type Htn1423_nghiasanphamsstatuscode = keyof typeof Htn1423_nghiasanphamsstatuscode;

export interface Htn1423_nghiasanphamsBase {
  htn1423_name: string;
  htn1423_nghiasanphamid: string;
  "htn1423_Nhomsanpham@odata.bind": string;
  htn1423_onvichuan: string;
  htn1423_quycach: string;
  htn1423_thuonghieu: string;
  importsequencenumber?: string;
  overriddencreatedon?: string;
  ownerid: string;
  owneridtype: string;
  statecode: Htn1423_nghiasanphamsstatecode;
  statuscode?: Htn1423_nghiasanphamsstatuscode;
  timezoneruleversionnumber?: string;
  utcconversiontimezonecode?: string;
}

export interface Htn1423_nghiasanphams extends Htn1423_nghiasanphamsBase {
  createdbyname?: string;
  createdbyyominame: string;
  createdon?: string;
  createdonbehalfbyname?: string;
  createdonbehalfbyyominame: string;
  htn1423_nhomsanphamname?: string;
  htn1423_soonhangban?: string;
  htn1423_soonhangban_date?: string;
  htn1423_soonhangban_state?: string;
  htn1423_soonhangmua?: string;
  htn1423_soonhangmua_date?: string;
  htn1423_soonhangmua_state?: string;
  htn1423_tenayusanpham?: string;
  modifiedbyname?: string;
  modifiedbyyominame: string;
  modifiedon?: string;
  modifiedonbehalfbyname?: string;
  modifiedonbehalfbyyominame: string;
  owneridname: string;
  owneridyominame: string;
  owningbusinessunitname: string;
  statecodename?: string;
  statuscodename?: string;
  versionnumber?: string;
  createdby?: object;
  _createdby_value?: string;
  createdonbehalfby?: object;
  _createdonbehalfby_value?: string;
  htn1423_nhomsanpham?: object;
  _htn1423_nhomsanpham_value?: string;
  modifiedby?: object;
  _modifiedby_value?: string;
  modifiedonbehalfby?: object;
  _modifiedonbehalfby_value?: string;
  owningbusinessunit?: object;
  _owningbusinessunit_value?: string;
  owningteam?: object;
  _owningteam_value?: string;
  owninguser?: object;
  _owninguser_value?: string;
}